+++ zcm/lib/helpers.py
@@ -13,8 +13,12 @@
 # limitations under the License.
 
 import argparse
+import logging
 import random
 import string
+import subprocess
+
+log = logging.getLogger(__name__)
 
 
 def check_positive(value, min_value=0):
@@ -24,6 +28,7 @@ def check_positive(value, min_value=0):
             "%s is an invalid positive int value" % value)
     return ivalue
 
+
 def check_one_or_more(value):
     return check_positive(value, 1)
 
@@ -31,3 +36,26 @@ def check_one_or_more(value):
 # https://stackoverflow.com/questions/2257441/random-string-generation-with-upper-case-letters-and-digits
 def id_generator(size=10, chars=string.ascii_uppercase + string.digits):
     return ''.join(random.choice(chars) for _ in range(size))
+
+
+# shutil.copytree does not copy owner
+def copy_directory(source, target):
+    if not source.is_dir() or not target.is_dir():
+        return -1
+    cmd1 = ['tar', 'cf', '-', '.']
+    cmd2 = ['tar', 'xf', '-']
+    log.debug('Running command: "cd %s; %s | (cd %s; %s)"' % (str(source), ' '.join(cmd1), str(target), ' '.join(cmd2)))
+    process1 = subprocess.Popen(cmd1, stdout=subprocess.PIPE, cwd=source)
+    process2 = subprocess.Popen(cmd2, stdin=process1.stdout, cwd=target)
+    process1.stdout.close()
+    stdout, stderr = process2.communicate()
+    if stdout:
+        for line in iter(stdout.splitlines()):
+            log.info(line)
+    if stderr:
+        for line in iter(stderr.splitlines()):
+            if process2.returncode == 0:
+                log.warning(line)
+            else:
+                log.error(line)
+    return process2.returncode